+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+### Changed
+- Switch build tools to [pixi].
+- Upgraded [zensols.deeplearn] to 1.14.2.
+
+
 ## [0.0.1] - 2025-05-04
 ### Added
 - Initial version.
@@ -16,3 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- links -->
 [Unreleased]: https://github.com/plandes/lmtask/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/plandes/lmtask/compare/v0.0.0...v0.0.1
+
+[pixi]: https://pixi.sh
+[zensols.deeplearn]: https://github.com/plandes/deeplearn